Splitting out ParallelTest.py
