+++ tests/SpiffWorkflow/bpmn/ParallelTest.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 
-
-
 from builtins import range
 import unittest
 import logging
@@ -11,511 +9,6 @@ from SpiffWorkflow.bpmn.workflow import BpmnWorkflow
 from tests.SpiffWorkflow.bpmn.BpmnWorkflowTestCase import BpmnWorkflowTestCase
 
 __author__ = 'matth'
-
-
-class ParallelFromCamunda(BpmnWorkflowTestCase):
-
-    # Should we move this to the Camunda package?  Is this even testing anything Camunda related?
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec('Test-Workflows/Parallel.camunda.bpmn20.xml', 'Process_1hb021r')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-    def testRunThroughParallelTaskFirst(self):
-
-        # 1 first task
-        self.assertEqual(1, len(self.workflow.get_tasks(TaskState.READY)))
-        self.do_next_named_step('First Task')
-        self.save_restore()
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-
-        # 3 parallel tasks
-        self.assertEqual(3, len(self.workflow.get_tasks(TaskState.READY)))
-        self.do_next_named_step('Parallel Task A')
-        self.save_restore()
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Parallel Task B')
-        self.save_restore()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Parallel Task C')
-        self.save_restore()
-        self.workflow.do_engine_steps()
-        self.save_restore()
-
-        # 1 last task
-        self.assertEqual(1, len(self.workflow.get_tasks(TaskState.READY)))
-        self.do_next_named_step('Last Task')
-        self.save_restore()
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-
-    def testAllParallelDataMakesItIntoGatewayTask(self):
-        """It should be true that data collected across parallel tasks
-        is all available in the join task."""
-
-        self.do_next_named_step('First Task')
-        self.do_next_named_step('Parallel Task A',
-                                set_attribs={"taskA": "taskA"})
-        self.do_next_named_step('Parallel Task B',
-                                set_attribs={"taskB": "taskB"})
-        self.do_next_named_step('Parallel Task C',
-                                set_attribs={"taskC": "taskC"})
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Last Task')
-        self.assertEquals("taskA", self.workflow.last_task.data["taskA"])
-        self.assertEquals("taskB", self.workflow.last_task.data["taskB"])
-        self.assertEquals("taskC", self.workflow.last_task.data["taskC"])
-
-
-class ParallelJoinLongInclusiveTest(BpmnWorkflowTestCase):
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec('Test-Workflows/Parallel-Join-Long-Inclusive.bpmn20.xml', 'Parallel Join Long Inclusive')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-    def testRunThroughThread1FirstThenNo(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step(
-            'Thread 1 - Choose', choice='Yes', with_save_load=True)
-        self.workflow.do_engine_steps()
-        for i in range(1, 13):
-            self.do_next_named_step('Thread 1 - Task %d' % i)
-            self.workflow.do_engine_steps()
-
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.assertEqual(1, len(self.workflow.get_tasks(TaskState.WAITING)))
-
-        self.do_next_named_step(
-            'Thread 2 - Choose', choice='No', with_save_load=True)
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Done', with_save_load=True)
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Thread 2 - No Task', with_save_load=True)
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testNoFirstThenThread1(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step(
-            'Thread 2 - Choose', choice='No', with_save_load=True)
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step(
-            'Thread 1 - Choose', choice='Yes', with_save_load=True)
-        self.workflow.do_engine_steps()
-        for i in range(1, 13):
-            self.do_next_named_step('Thread 1 - Task %d' % i)
-            self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done', with_save_load=True)
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Thread 2 - No Task', with_save_load=True)
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-
-class ParallelJoinLongTest(BpmnWorkflowTestCase):
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec('Test-Workflows/Parallel-Join-Long.bpmn20.xml', 'Parallel Join Long')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-    def testRunThroughAlternating(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step(
-            'Thread 1 - Choose', choice='Yes', with_save_load=True)
-        self.workflow.do_engine_steps()
-        self.do_next_named_step(
-            'Thread 2 - Choose', choice='Yes', with_save_load=True)
-        self.workflow.do_engine_steps()
-
-        for i in range(1, 13):
-            self.do_next_named_step(
-                'Thread 1 - Task %d' % i, with_save_load=True)
-            self.workflow.do_engine_steps()
-            self.do_next_named_step(
-                'Thread 2 - Task %d' % i, with_save_load=True)
-            self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done', with_save_load=True)
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testRunThroughThread1First(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step(
-            'Thread 1 - Choose', choice='Yes', with_save_load=True)
-        self.workflow.do_engine_steps()
-        for i in range(1, 13):
-            self.do_next_named_step('Thread 1 - Task %d' % i)
-            self.workflow.do_engine_steps()
-
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.assertEqual(1, len(self.workflow.get_tasks(TaskState.WAITING)))
-
-        self.do_next_named_step(
-            'Thread 2 - Choose', choice='Yes', with_save_load=True)
-        self.workflow.do_engine_steps()
-        for i in range(1, 13):
-            self.do_next_named_step(
-                'Thread 2 - Task %d' % i, with_save_load=True)
-            self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done', with_save_load=True)
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-
-class ParallelMultipleSplitsTest(BpmnWorkflowTestCase):
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec('Test-Workflows/Parallel-Multiple-Splits.bpmn20.xml', 'Parallel Multiple Splits')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-    def testRunThroughAlternating(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Do First')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('SP 1 - Choose', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('SP 2 - Choose', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('SP 3 - Choose', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('SP 1 - Yes Task')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('SP 2 - Yes Task')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('SP 3 - Yes Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-
-class ParallelOnePathEndsTest(BpmnWorkflowTestCase):
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec('Test-Workflows/Parallel-One-Path-Ends.bpmn20.xml', 'Parallel One Path Ends')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-    def testRunThroughParallelTaskFirst(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Parallel Task')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Choice 1', choice='No')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testRunThroughChoiceFirst(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Choice 1', choice='No')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Parallel Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testRunThroughParallelTaskFirstYes(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Parallel Task')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Choice 1', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Yes Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-
-class ParallelThenExclusiveTest(BpmnWorkflowTestCase):
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec('Test-Workflows/Parallel-Then-Exclusive.bpmn20.xml', 'Parallel Then Exclusive')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-    def testRunThroughParallelTaskFirst(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Parallel Task')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Choice 1', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Yes Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testRunThroughChoiceFirst(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Choice 1', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Parallel Task')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Yes Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testRunThroughChoiceThreadCompleteFirst(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Choice 1', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Yes Task')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Parallel Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-
-class ParallelThenExclusiveNoInclusiveTest(ParallelThenExclusiveTest):
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec(
-            'Test-Workflows/Parallel-Then-Exclusive-No-Inclusive.bpmn20.xml', 
-            'Parallel Then Exclusive No Inclusive')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-class ParallelThroughSameTaskTest(BpmnWorkflowTestCase):
-
-    def setUp(self):
-        spec, subprocesses = self.load_workflow_spec('Test-Workflows/Parallel-Through-Same-Task.bpmn20.xml', 'Parallel Through Same Task')
-        self.workflow = BpmnWorkflow(spec, subprocesses)
-        self.workflow.do_engine_steps()
-
-    def testRunThroughFirstRepeatTaskFirst(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Repeated Task')
-        self.workflow.do_engine_steps()
-        # The inclusive gateway allows this to pass through (since there is a
-        # route to it on the same sequence flow)
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Choice 1', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Yes Task')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Repeated Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testRepeatTasksReadyTogether(self):
-
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Choice 1', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Yes Task')
-        self.workflow.do_engine_steps()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        ready_tasks = self.workflow.get_tasks(TaskState.READY)
-        self.assertEqual(2, len(ready_tasks))
-        self.assertEqual(
-            'Repeated Task', ready_tasks[0].task_spec.description)
-        ready_tasks[0].complete()
-        self.workflow.do_engine_steps()
-        # The inclusive gateway allows us through here, because there is no route for the other thread
-        # that doesn't use the same sequence flow
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.do_next_named_step('Repeated Task')
-        self.workflow.do_engine_steps()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testRepeatTasksReadyTogetherSaveRestore(self):
-
-        self.save_restore()
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Choice 1', choice='Yes')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('Yes Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        ready_tasks = self.workflow.get_tasks(TaskState.READY)
-        self.assertEqual(2, len(ready_tasks))
-        self.assertEqual(
-            'Repeated Task', ready_tasks[0].task_spec.description)
-        ready_tasks[0].complete()
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        # The inclusive gateway allows us through here, because there is no route for the other thread
-        # that doesn't use the same sequence flow
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('Repeated Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testNoRouteRepeatTaskFirst(self):
-
-        self.save_restore()
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Repeated Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        # The inclusive gateway allows this to pass through (since there is a
-        # route to it on the same sequence flow)
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('Choice 1', choice='No')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('No Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testNoRouteNoTaskFirst(self):
-
-        self.save_restore()
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Choice 1', choice='No')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('No Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Repeated Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
-    def testNoRouteNoFirstThenRepeating(self):
-
-        self.save_restore()
-        self.assertEqual(2, len(self.workflow.get_tasks(TaskState.READY)))
-
-        self.do_next_named_step('Choice 1', choice='No')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('Repeated Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.assertRaises(AssertionError, self.do_next_named_step, 'Done')
-        self.do_next_named_step('No Task')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-        self.do_next_named_step('Done')
-        self.workflow.do_engine_steps()
-        self.save_restore()
-
-        self.assertEqual(
-            0, len(self.workflow.get_tasks(TaskState.READY | TaskState.WAITING)))
-
 
 class BaseParallelTest(BpmnWorkflowTestCase):
